Add and refactor new abstract class TimedTask
A TimedTask is a task that has a time parameter.
