+++ src/main/java/luoyi/duke/data/task/Task.java
@@ -4,7 +4,7 @@ package luoyi.duke.data.task;
  * Immutable Task Object.
  * A task has a description and a completion status.
  */
-public class Task implements ITask {
+public abstract class Task implements ITask {
     protected final String description;
     protected final boolean isDone;
 
@@ -16,24 +16,6 @@ public class Task implements ITask {
     protected Task(String description, boolean isDone) {
         this.description = description;
         this.isDone = isDone;
-    }
-
-    /**
-     * Returns a new uncompleted task.
-     *
-     * @param description Description of task.
-     * @return New uncompleted task.
-     */
-    public static Task getTask(String description) {
-        return new Task(description, false);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public Task markComplete() {
-        return new Task(this.description, true);
     }
 
     /**
@@ -58,10 +40,6 @@ public class Task implements ITask {
     @Override
     public String getDescription() {
         return description;
-    }
-
-    public boolean isSameTime(String time) {
-        return false;
     }
 
     @Override
