+++ src/main/java/luoyi/duke/data/task/ToDo.java
@@ -1,5 +1,9 @@
 package luoyi.duke.data.task;
 
+import java.util.Optional;
+
+import luoyi.duke.common.TimeWrapper;
+
 /**
  * ToDo class.
  * A ToDo is a task without any date/time attached to it.
@@ -32,6 +36,27 @@ public class ToDo extends Task {
     @Override
     public String getDataString() {
         return String.format("T|%d|%s", isDone ? 1 : 0, description);
+    }
+
+    /**
+     * Returns an empty optional.
+     * This is because todo has no time.
+     *
+     * @return Empty optional.
+     */
+    @Override
+    public Optional<TimeWrapper> getTime() {
+        return Optional.empty();
+    }
+
+    /**
+     * Returns vacuous false.
+     *
+     * @return False.
+     */
+    @Override
+    public boolean isSameTime(String date) {
+        return false;
     }
 
     @Override
